Move diff & fraction utilities to global library.
