+++ packages/field-kit/src/utils/index.js
@@ -1,5 +1,7 @@
 // Only these utilities are exposed to field modules on the window.lib global namespace.
+export { default as daysAway } from './daysAway';
+export { default as diff } from './diff';
+export { gcd, toDecimal, toFraction } from './fraction';
 export { mergeGeometries, removeGeometry, isNearby } from './geometry';
 export { default as parseNotes } from './parseNotes';
-export { default as daysAway } from './daysAway';
 export { default as parseFilter } from './parseFilter';
