+++ packages/field-module-tasks/src/components/Tasks.vue
@@ -2,31 +2,11 @@
   <router-view/>
 </template>
 <script>
-const { useEntities } = window.lib;
+const { toFraction, useEntities } = window.lib;
 const {
   computed, ref, inject, provide,
 } = window.Vue;
 const { useRouter } = window.VueRouter;
-
-const gcd = (a, b) => (b > Number.EPSILON ? gcd(b, a % b) : a);
-function toFraction(string) {
-  const decimal = Number.parseFloat(string);
-  if (Number.isInteger(decimal)) {
-    return {
-      numerator: decimal,
-      denominator: 1,
-      decimal,
-    };
-  }
-  const int = Number.parseInt(string, 10);
-  const places = (decimal - int).toString(10).length - 2;
-  let denominator = Math.abs(10 ** places);
-  let numerator = decimal * denominator;
-  const divisor = gcd(numerator, denominator);
-  denominator /= divisor;
-  numerator /= divisor;
-  return { numerator, denominator, decimal };
-}
 
 export default {
   name: 'TasksContainer',
